Refactor Unit Tests into their own file / update README
